Explicitly note obsolete/unmaintained fields.
A number of fields are documented as being unmaintained, so the value
will always be zero. Previously two of these were noted as "zero1" and
"zero2", which isn't very useful. This change gives them much more
meaningful names "obs_<field-name>" (so "obs_nswap" and "obs_cnswap"),
as well as updating the now-obsolete "it_real_value" field to
"obs_itrealvalue".